+++ vitest.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
         }
     },
     test: {
-        include: ["src/**/*.test.ts", "packages/**/*.test.ts"],
+        include: ["src/**/*.test.ts", "packages/**/*.test.ts", "scripts/**/*.test.ts"],
         environment: "node",
         globals: true
     }